perf: lazy load Message Header so non auth is smaller
