+++ packages/client/bundlesize.config.js
@@ -5,16 +5,16 @@ export default {
 	},
 	sizes: [
 		{
-			path: "dist/index.html",
-			limit: "2 kb",
-		},
-		{
 			path: "dist/assets/index-<hash>.js",
-			limit: "25 kb",
+			limit: "2 kb",
 		},
 		{
 			path: "dist/assets/index-<hash>.css",
 			limit: "10 kb",
+		},
+		{
+			path: "dist/LazyHeader-<hash>.js",
+			limit: "12 kb",
 		},
 		{
 			path: "dist/MessageAssistant-<hash>.js",
